[Refactor] 디렉토리 구조 개선
애플리케이션의 가독성을 위해 디렉토리 구조를 개선했습니다.

- http, request/response 디렉토리를 추가하여 HTTP 요청 및 응답 처리 관련 클래스들을 별도로 분리
