+++ src/main/java/utils/FileUtil.java
@@ -1,8 +1,7 @@
-package util;
+package utils;
 
 import java.io.File;
 import java.io.IOException;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 
 public class FileUtil {
