+++ src/main/java/webserver/http/response/HttpResponse.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.http.response;
 
 import java.util.HashMap;
 import java.util.Map;
